fix: extract watch.rs tests to sibling file (962→749 LOC)
Move #[cfg(test)] mod tests from watch.rs to watch_tests.rs via
#[path] declaration. Fixes file_size_invariant (max 750 LOC).


Agend-Agent: claude-aef7c0
Agend-Task: t-20260718013438626159-52837-1
Agend-Branch: feat/arch14-ci-handoff-task-defer
Agend-Issued-At: 2026-07-18T02:44:31.059128+00:00
